Module llamapun::ams

Representation, normalization and utilities for working with AMS markup in LaTeX-derived scientific documents

Enums

AmsEnv

Author-annotated \newthorem{} environments using the amsthm.sty mechanism which are then transported in the HTML representation as ltx_theorem_<env>

StructuralEnv

Semantically fixed structural environments in scientific documents, to collect as add-on to the AMS markup

Functions

class_to_env

Maps a latexml-produced HTML class, such as “ltx_theorem ltx_theorem_lemma” to an AmsEnv enum

has_markup

Checks a llamapun Document for ‘ltx_theorem’ AMS markup

has_markup_xmldoc

Checks a libxml document for ltx_theorem AMS markup

normalize_env

If known, maps a commonly used AMS environment to a shortlist of the 23 most notable environments. Returns None if there is no AMS markup. Most experiments would be best off with dropping the AmsEnv::Other resulting paragraphs, to avoid unintended dilution of the known environments.