Enum llamapun::ams::AmsEnv

pub enum AmsEnv {
    Acknowledgement,
    Algorithm,
    Answer,
    Affirmation,
    Assumption,
    Bound,
    Caption,
    Case,
    Claim,
    Comment,
    Conclusion,
    Condition,
    Conjecture,
    Constraint,
    Convention,
    Corollary,
    Criterion,
    Definition,
    Demonstration,
    Discussion,
    Example,
    Experiment,
    Expansion,
    Expectation,
    Explanation,
    Fact,
    Hint,
    Issue,
    Keywords,
    Lemma,
    Notation,
    Note,
    Notice,
    Observation,
    Paragraph,
    Principle,
    Problem,
    Proof,
    Proposition,
    Question,
    Remark,
    Result,
    Rule,
    Solution,
    Step,
    Summary,
    Theorem,
    Other,
}

Author-annotated \newthorem{} environments using the amsthm.sty mechanism which are then transported in the HTML representation as ltx_theorem_<env>

We shortlist 23 of the >20,000 variety of values present in the arXiv corpus

Variants

Acknowledgement

typically co-author support for a proof/paper (also "thanks")

Algorithm

usually defines a computer science algorithm (also "heuristic"; arXiv data is bad quality, would not recommend using)

Answer

To be analyzed (?)

Affirmation

To be analyzed (?)

Assumption

assumption/axiom/assertion/prior -- should they be included in propositions? Are they separable?

Bound

To be analyzed (?)

Caption

usually an actual Figure or Table captions realized via AMS (strange data, would not recommend using)

Case

A case in a multi-step proof / description / exposition

Claim

To be analyzed (?)

Comment

To be analyzed (?)

Conclusion

To be analyzed (?)

Condition

Potentially a constraint on a proof

Conjecture

An unproven statement/theorem (includes "conjecture", "ansatz", "guess", "hypothesis")

Constraint

To be analyzed (?)

Convention

To be analyzed (?)

Corollary

A direct-to-derive consequence of a prior proposition

Criterion

To be analyzed (?)

Definition

Unlike notations, introduces new conceptual mathematical objects

Demonstration

To be analyzed (?)

Discussion

To be analyzed (?)

Example

Demonstration of a definition, notation etc (also "experiment")

Experiment

To be analyzed (?)

Expansion

To be analyzed (?)

Expectation

To be analyzed (?)

Explanation

To be analyzed (?)

Fact

To be analyzed (?)

Hint

To be analyzed (?)

Issue

To be analyzed (?)

Keywords

To be analyzed (?)

Lemma

A smaller sub-theorem to a main theorem

Notation

Introduces a new syntactic rule, usually for convenience / brevity

Note

To be analyzed (?)

Notice

To be analyzed (?)

Observation

To be analyzed (?)

Paragraph

A named paragraph, without a clear standalone function

Principle

To be analyzed (?)

Problem

A task to be solved (sometimes with solution following), includes "Exercise"

Proof

Proves a prior theorem/lemma

Proposition

A provably true/false statement. Is this a synonym to theorem in arXiv?

Question

(sometimes) initial goal of inquiry (also "puzzle", "query")

Remark

A comment that is an aside to the main line of reasoning

Result

Summarizes paper's experimental deliverables

Rule

To be analyzed (?)

Solution

To be analyzed (?)

Step

A part of a proof, or demonstration/layout

Summary

To be analyzed (?)

Theorem

A main proposition to be proven in the document

Other

Anything else that was marked up with AMS, but doesn't fit this scheme