Enum llamapun::ams::StructuralEnv

pub enum StructuralEnv {

    Abstract,
    Acknowledgement,
    Analysis,
    Application,
    Assumption,
    Background,
    Case,
    Caption,
    Claim,
    Conclusion,
    Condition,
    Conjecture,
    Contribution,
    Corollary,
    Data,
    Dataset,
    Definition,
    Demonstration,
    Description,
    Discussion,
    Example,
    Experiment,
    Fact,
    FutureWork,
    Implementation,
    Introduction,
    Lemma,
    Methods,
    Model,
    Motivation,
    Notation,
    Observation,
    Other,
    Preliminaries,
    Problem,
    Proof,
    Property,
    Proposition,
    Question,
    RelatedWork,
    Remark,
    Result,
    Simulation,
    Step,
    Summary,
    Theorem,
    Theory,

}

Semantically fixed structural environments in scientific documents, to collect as add-on to the AMS markup

Note we are explicitly ignoring some of the very high-frequency environments, as they are not rich on textual content. Namely: references, appendix, pacs, subject; Which are rich in metadata and semi-structured content (figures, tables).

Variants

Abstract

Acknowledgement

Analysis

Application

Assumption

Background

Case

Caption

Claim

Conclusion

Condition

Conjecture

Contribution

Corollary

Data

Dataset

Definition

Demonstration

Description

Discussion

Example

Experiment

Fact

FutureWork

Implementation

Introduction

Lemma

Methods

Model

Motivation

Notation

Observation

Other

Preliminaries

Problem

A task to be solved (sometimes with solution following), includes “Exercise”

Proof

Property

Proposition

Question

RelatedWork

Remark

Result

Simulation

Step

Summary

Theorem

Theory