Struct llamapun::data::Document

pub struct Document<'d> {
    pub dom: XmlDoc,
    pub path: String,
    pub corpus: &'d Corpus,
    pub dnm: Option<DNM>,
}

One of our math documents.

Fields
dom: XmlDoc

The DOM of the document

path: String

The file path of the document

corpus: &'d Corpus

A reference to the corpus containing this document

dnm: Option<DNM>

If it exists, the DNM corresponding to this document

Methods

impl<'d> Document<'d>

fn new(filepath: String, owner: &'d Corpus) -> Result<Self, XmlParseError>

Load a new document

fn paragraph_iter(&mut self) -> ParagraphIterator

Get an iterator over the paragraphs of the document

fn sentence_iter(&mut self) -> SentenceIterator

Get an iterator over the sentences of the document