Struct llamapun::dnm::DNMParameters

pub struct DNMParameters {
    pub special_tag_name_options: HashMap<String, SpecialTagsOption>,
    pub special_tag_class_options: HashMap<String, SpecialTagsOption>,
    pub normalize_white_spaces: bool,
    pub wrap_tokens: bool,
    pub move_whitespaces_between_nodes: bool,
    pub normalize_unicode: bool,
    pub stem_words_once: bool,
    pub stem_words_full: bool,
    pub convert_to_lowercase: bool,
}

Paremeters for the DNM generation

Fields
special_tag_name_options: HashMap<String, SpecialTagsOption>

How to deal with special tags (e.g. <math> tags)

special_tag_class_options: HashMap<String, SpecialTagsOption>

How to deal with tags with special class names (e.g. ltx_note_mark) Remark: If both a tag name and a tag class match, the tag name rule will be applied.

normalize_white_spaces: bool

merge sequences of whitespaces into a single ' '. Doesn't affect tokens

wrap_tokens: bool

put spaces before and after tokens

move_whitespaces_between_nodes: bool

if there is a trailing white space in a tag, don't make it part of that tag. Requires normalize_white_spaces to be set.

normalize_unicode: bool

Replace unicode characters by the ascii code representation

stem_words_once: bool

Apply the morpha stemmer once to the text nodes

stem_words_full: bool

Apply the morpha stemmer to the text nodes as often as it changes something

convert_to_lowercase: bool

Move to lowercase (remark: The stemmer does automatically)

Methods

impl DNMParameters

fn llamapun_normalization() -> DNMParameters

Normalize in a reasonable way for our math documents

fn check(&self)

Prints warnings, if the parameter settings don't make sense. Doesn't check for every possible stupidity