Module llamapun::parallel_data

Data structures and Iterators for rayon-enabled parallel processing including parallel I/O in walking a corpus as well as DOM primitives that allow parallel iterators on XPath results, etc

Re-exports

pub use self::corpus::Corpus;

pub use self::document::Document;

Modules

corpus
container and API for a Corpus capable of parallel walks over its documents

document
container and API for a Document yielded during a parallel corpus walk

Structs

DNMRangeIterator
A generic iterator over DNMRanges with their associated document (e.g. for sentences)

ItemDNM
A DNM with associated document parent (e.g. for paragraphs, headings)

ItemDNMRange
A DNMRange with associated document

RoNodeIterator
Generic iterater over read-only xml nodes. It is the responsibility of the abstraction returning NodeIterator to specify the grouping principle for collecting the nodes

Traits

XPathFilteredIterator
An iterator adaptor for filtered selections over a document