Trait llamapun::parallel_data::XPathFilteredIterator

pub trait XPathFilteredIterator<'p> {
    fn to_sentences(&'p self) -> Vec<DNMRange<'p>>;
    fn get_document(&'p self) -> &'p Document<'_>;

    fn iter(&'p mut self) -> DNMRangeIterator<'p> { ... }
}

An iterator adaptor for filtered selections over a document

Required Methods

fn to_sentences(&'p self) -> Vec<DNMRange<'p>>

the sentences for the resulting selection

fn get_document(&'p self) -> &'p Document<'_>

the owner document being selected over

Provided Methods

fn iter(&'p mut self) -> DNMRangeIterator<'p>

Get an iterator over the sentences in this paragraph