Module llamapun::patterns

A module for pattern matching in mathematical documents

Structs

Match

A Match. Note that matches are represented in a tree structure, following the structure of the patterns

MathMarker

A marked math node

PatternFile

Contains rules loaded from a pattern file

PatternMarker

The marker used for marking patterns. If a match was found, this marker will be returned (along with the marked range).

TextMarker

A marked text range

Enums

MarkerEnum

Any marked result

Functions

match_sentence

returns the matches in a sentence