Module llamapun::util::path_helpers

Helpers intended mostly for non-Rust use, where rust is passed in a filesystem path and a llamapun-supported processing is requested

Functions

path_to_words
Given a path to a document, return a word-tokenized string of all of its paragraphs