Crate llamapun

The LLaMaPUn library in Rust

Language and Mathematics Processing and Understanding Common data structures and algorithms for semi-structured NLP on math-rich documents.

Modules

data
Data structures and Iterators for convenient high-level syntax

dnm
The dnm can be used for easier switching between the DOM (Document Object Model) representation and the plain text representation, which is needed for most NLP tools.

ngrams
A small ngram library ngrams are sequences of n consecutive words

stopwords
A tiny stopwords library Stopwords are words frequent words like "the", "it", "then", which would add too much noise to certain statistical methods

tokenizer
Provides functionality for tokenizing sentences and words

util
Various useful code snippets

Macros

print_stderr

println_stderr

record_node_map