Module llamapun::data

Data structures and Iterators for convenient high-level syntax

Structs

Corpus

An iterable Corpus of HTML5 documents

Document

One of our math documents.

DocumentIterator

File-system iterator yielding individual documents

Paragraph

A paragraph of a document with a DNM

ParagraphIterator

An iterator over paragraphs of a Document. Ignores paragraphs containing ltx_ERROR markup

SennaWordIterator

An iterator over the words of a sentence, where the words (and potentially additional information) are obtained using senna

Sentence

A sentence in a document

SentenceIterator

An iterator over the sentences of a document/paragraph

SimpleWordIterator

An iterator over the words of a sentence, where the words are only defined by their ranges

Word

A word with a POS tag