Module llamapun::dnm

The dnm can be used for easier switching between the DOM (Document Object Model) representation and the plain text representation, which is needed for most NLP tools.

Modules

node

Node auxiliaries for DNMs

Structs

DNM

The DNM is essentially a wrapper around the plain text representation of the document, which facilitates mapping plaintext pieces to the DOM. This breaks, if the DOM is changed after the DNM generation!

DNMParameters

Parameters for the DNM generation

DNMRange

Very often we’ll talk about substrings of the plaintext - words, sentences, etc. A DNMRange stores start and end point of such a substring and has a reference to the DNM.

RuntimeParseData

Some temporary data for the parser

Enums

SpecialTagsOption

Specifies how to deal with a certain tag