Struct llamapun::dnm::DNM

pub struct DNM {
    pub plaintext: String,
    pub parameters: DNMParameters,
    pub root_node: Node,
    pub node_map: HashMap<usize, (usize, usize)>,
    pub runtime: RuntimeParseData,
}

The DNM is essentially a wrapper around the plain text representation of the document, which facilitates mapping plaintext pieces to the DOM. This breaks, if the DOM is changed after the DNM generation!

Fields
plaintext: String

The plaintext

parameters: DNMParameters

The options for generation

root_node: Node

The root node of the underlying xml tree

node_map: HashMap<usize, (usize, usize)>

Maps nodes to plaintext offsets

runtime: RuntimeParseData

A runtime object used for holding auxiliary state

Methods

impl DNM

fn to_c14n_basic(&self) -> String

Our linguistic canonical form will only include 1) node name, 2) class attribute and 3) textual content - excludes certain experimental markup, such as all math annotation elements - excludes whitespace nodes and comment nodes

fn node_c14n_basic(&self, node: &Node) -> String

Canonicalize a single node of choice

fn to_hash_basic(&self) -> String

Obtain an MD5 hash from the canonical string of the entire DOM

fn node_hash_basic(&self, node: &Node) -> String

Obtain an MD5 hash from the canonical string of a Node

impl DNM

fn new(root: Node, parameters: DNMParameters) -> DNM

Creates a DNM for root

fn get_range_of_node(&self, node: &Node) -> Result<DNMRange, ()>

Get the plaintext range of a node